Moved mixnode status route to node status api module
